Old 'Button' is deprecated. Refactor 'DefaultButton' into new 'Button'
