+++ labsystem/src/Button/Button.js
@@ -2,12 +2,17 @@ import React from "react";
 import PropTypes from "prop-types";
 import AbstractButton from "./AbstractButton";
 
-export default class TextButton extends React.Component {
+export default class Button extends React.Component {
   static propTypes = {
     text: PropTypes.string.isRequired,
     skin: PropTypes.oneOf([
       "light",
-      "dark"
+      "dark",
+      "warning",
+      "destructive",
+      "warning-invert",
+      "destructive-invert",
+      "confirmation-invert",
     ]),
     icon: PropTypes.string,
     size: PropTypes.oneOf(["normal", "small", "large"]),
@@ -17,11 +22,12 @@ export default class TextButton extends React.Component {
   static defaultProps = {
     skin: undefined,
     icon: undefined,
+    size: "normal",
     disabled: false,
   };
 
   render() {
     // eslint-disable-next-line react/jsx-props-no-spreading
-    return <AbstractButton variant="text" {...this.props} />;
+    return <AbstractButton variant="default" {...this.props} />;
   }
 }
