+++ labsystem/src/Button/OutlineButton.js
@@ -2,18 +2,10 @@ import React from "react";
 import PropTypes from "prop-types";
 import AbstractButton from "./AbstractButton";
 
-export default class DefaultButton extends React.Component {
+export default class OutlineButton extends React.Component {
   static propTypes = {
     text: PropTypes.string.isRequired,
-    skin: PropTypes.oneOf([
-      "light",
-      "dark",
-      "warning",
-      "destructive",
-      "warning-invert",
-      "destructive-invert",
-      "confirmation-invert",
-    ]),
+    skin: PropTypes.oneOf(["light", "dark"]),
     icon: PropTypes.string,
     size: PropTypes.oneOf(["normal", "small", "large"]),
     disabled: PropTypes.bool,
@@ -27,6 +19,7 @@ export default class DefaultButton extends React.Component {
   };
 
   render() {
-    return <AbstractButton variant="default" {...this.props} />;
+    // eslint-disable-next-line react/jsx-props-no-spreading
+    return <AbstractButton variant="outline" {...this.props} />;
   }
 }
