+++ labsystem/src/Button/TextButton.js
@@ -2,13 +2,10 @@ import React from "react";
 import PropTypes from "prop-types";
 import AbstractButton from "./AbstractButton";
 
-export default class OutlineButton extends React.Component {
+export default class TextButton extends React.Component {
   static propTypes = {
     text: PropTypes.string.isRequired,
-    skin: PropTypes.oneOf([
-      "light",
-      "dark"
-    ]),
+    skin: PropTypes.oneOf(["light", "dark"]),
     icon: PropTypes.string,
     size: PropTypes.oneOf(["normal", "small", "large"]),
     disabled: PropTypes.bool,
@@ -17,11 +14,12 @@ export default class OutlineButton extends React.Component {
   static defaultProps = {
     skin: undefined,
     icon: undefined,
+    size: "normal",
     disabled: false,
   };
 
   render() {
     // eslint-disable-next-line react/jsx-props-no-spreading
-    return <AbstractButton variant="outline" {...this.props} />;
+    return <AbstractButton variant="text" {...this.props} />;
   }
 }
